proninteam/requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.3
django-choices==1.7.2
django-cors-headers==3.13.0
django-pytest==0.2.0
djangorestframework==3.14.0
exceptiongroup==1.0.4
flake8==5.0.4
gunicorn==20.0.4
importlib-metadata==4.2.0
iniconfig==1.1.1
mccabe==0.7.0
packaging==21.3
pillow==9.3.0
pluggy==1.0.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.4.0
zipp==3.10.0
